Rough backend draft
Can now access a mongo database. Contains rough drafts for database data structures. Connects to a mongo db I created, to change, set the variable in the .env file to your target link. See ReadMe.txt and comments for more information on file structure and how to start a connection to a mongodb.